requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
numpy==1.24.3
librosa==0.10.1
soundfile==0.12.1
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
openai-whisper==20231117
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.1.3
scikit-learn==1.3.2
scipy==1.11.4
chromadb==0.4.22
sentence-transformers==2.7.0